// frontend/web/src/services/API/proserve-wb-packaging-api/models/GetRecipeResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Recipe } from './Recipe';
import {
    RecipeFromJSON,
    RecipeFromJSONTyped,
    RecipeToJSON,
} from './Recipe';

/**
 * 
 * @export
 * @interface GetRecipeResponse
 */
export interface GetRecipeResponse {
    /**
     * 
     * @type {Recipe}
     * @memberof GetRecipeResponse
     */
    recipe: Recipe;
}

/**
 * Check if a given object implements the GetRecipeResponse interface.
 */
export function instanceOfGetRecipeResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "recipe" in value;

    return isInstance;
}

export function GetRecipeResponseFromJSON(json: any): GetRecipeResponse {
    return GetRecipeResponseFromJSONTyped(json, false);
}

export function GetRecipeResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetRecipeResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'recipe': RecipeFromJSON(json['recipe']),
    };
}

export function GetRecipeResponseToJSON(value?: GetRecipeResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'recipe': RecipeToJSON(value.recipe),
    };
}